tracj code added

+++ track_points_trajectory.py
@@ -6,13 +6,17 @@ from math import radians
 import glob
 from pcd_generator import generate_pcd
 import pickle
+import math
 
 
 def get_traj(P1, P2, v_b, t, prev_point):
     C1 = (v_b * t) - P1[1] + (P1[0]**2 / (v_b * t - P1[1]))
     B1 = P2[0] + (P2[1] * P1[0] / (v_b * t - P1[1]))
     translation_magnitude = v_b*t #eclid(P1,P2)#v_b*t
-    angle = np.arcsin((-B1) / C1)
+    value = (-B1) / C1
+    if value > 1 or value < -1:
+        return None
+    angle = np.arcsin(value)
     rotation_matrix = np.array([
         [np.cos(angle), -np.sin(angle)],
         [np.sin(angle), np.cos(angle)]
@@ -32,9 +36,15 @@ def eclid(p1,p2):
 if __name__ == "__main__":
     files = glob.glob('../mmPhase/datasets/stick*_traj_cse_*.bin')
     for file in files:
-        pcd, velocity = generate_pcd_and_speed(file)
-        with open('data.pkl', 'wb') as f:
-            pickle.dump({'pcd': pcd, 'velocity': velocity}, f)
+        pickle_save_name = file.split('/')[-1].split('.')[0] + 'pcd_vb_data.pkl'
+        if os.path.exists(pickle_save_name):
+            with open(pickle_save_name, 'rb') as f:
+                data = pickle.load(f)
+            pcd, velocity = data['pcd'], data['velocity']
+        else:
+            pcd, velocity = generate_pcd_and_speed(file)
+            with open(pickle_save_name, 'wb') as f:
+                pickle.dump({'pcd': pcd, 'velocity': velocity}, f)
         print(pcd.shape)
         print(velocity.shape)
         plt.cla()
@@ -52,14 +62,20 @@ if __name__ == "__main__":
                 prev_pointclouds = points
                 continue
             for point1 in points:
+                if eclid(point1, (0,0)) < 0.05:
+                    continue
                 for point2 in prev_pointclouds:
+                    if eclid(point2, (0,0)) < 0.05: 
+                        continue
                     distance = eclid(point1, point2)
                     if distance < 0.05:
                         new_traj_point = get_traj(point1, point2,v_b, 0.2, prev_traj_point)
+                        if new_traj_point is None:
+                            continue
                         traj1.append(new_traj_point)
             if len(traj1) > 1:
                 final_traj_point = (np.median([point[0] for point in traj1]), np.median([point[1] for point in traj1]))
-                print(final_traj_point)
+                print(final_traj_point, v_b)
             else:
                 print("Skipped frame_no: ", frame_no)
             # print("final_traj_point: ", final_traj_point)
@@ -70,7 +86,7 @@ if __name__ == "__main__":
             
             
         plt.plot([e[0] for e in final_trajectory],[e[1] for e in final_trajectory])
-        save_fig = f"trajectory_plot_{traj_file.split('.')[0]}.png"
+        save_fig = f"trajectory_plot_{file.split('/')[-1].split('.')[0]}.png"
         plt.savefig(save_fig)
         plt.close()
         print(save_fig)
